Reorganize test suites for wave 6
